API: libcrmcommon: Deprecate mainloop_child_t API
Deprecate the mainloop_child_t type itself and all functions that accept
or manipulate mainloop_child_t objects. External programs should not use
Pacemaker for general-purpose asynchronous programming.
